unity-app/tests/ecu_simulator/requirements.txt
python-can>=4.0.0
can-isotp>=2.0.0